+++ src/internet/model/tcp-prr-recovery.h
@@ -50,15 +50,6 @@ class TcpPrrRecovery : public TcpClassicRecovery
 
     ~TcpPrrRecovery() override;
 
-    /**
-     * \brief Reduction Bound variant (CRB or SSRB)
-     */
-    enum ReductionBound_t
-    {
-        CRB, /**< Conservative Reduction Bound */
-        SSRB /**< Slow Start Reduction Bound */
-    };
-
     std::string GetName() const override;
 
     void EnterRecovery(Ptr<TcpSocketState> tcb,
@@ -66,7 +57,7 @@ class TcpPrrRecovery : public TcpClassicRecovery
                        uint32_t unAckDataCount,
                        uint32_t deliveredBytes) override;
 
-    void DoRecovery(Ptr<TcpSocketState> tcb, uint32_t deliveredBytes) override;
+    void DoRecovery(Ptr<TcpSocketState> tcb, uint32_t deliveredBytes, bool isDupAck) override;
 
     void ExitRecovery(Ptr<TcpSocketState> tcb) override;
 
@@ -78,7 +69,6 @@ class TcpPrrRecovery : public TcpClassicRecovery
     uint32_t m_prrDelivered{0};       //!< total bytes delivered during recovery phase
     uint32_t m_prrOut{0};             //!< total bytes sent during recovery phase
     uint32_t m_recoveryFlightSize{0}; //!< value of bytesInFlight at the start of recovery phase
-    ReductionBound_t m_reductionBoundMode{SSRB}; //!< mode of Reduction Bound to be used
 };
 } // namespace ns3