+++ src/internet/test/tcp-classic-recovery-test.cc
@@ -90,7 +90,7 @@ ClassicRecoveryTest::DoRun()
 
     uint32_t cWndInflPrevious = m_state->m_cWndInfl;
     uint32_t cWndPrevious = m_state->m_cWnd;
-    recovery->DoRecovery(m_state, 500);
+    recovery->DoRecovery(m_state, 500, false);
     NS_TEST_ASSERT_MSG_EQ(
         m_state->m_cWndInfl,
         (cWndInflPrevious + m_state->m_segmentSize),